httpserver/handler/debug_build_test.go
package handler_test

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/ankorstore/yokai/httpserver/handler"
	"github.com/labstack/echo/v4"
	"github.com/stretchr/testify/assert"
)

func TestDebugBuildHandler(t *testing.T) {
	t.Parallel()

	httpServer := echo.New()
	httpServer.GET("/debug/build", handler.DebugBuildHandler())

	req := httptest.NewRequest(http.MethodGet, "/debug/build", nil)
	rec := httptest.NewRecorder()
	httpServer.ServeHTTP(rec, req)

	assert.Equal(t, http.StatusOK, rec.Code)
	assert.Contains(t, rec.Body.String(), `"env":{`)
	assert.Contains(t, rec.Body.String(), `"go":{`)
}
